+++ modules/control-panel-mail/src/MailServiceProvider.php
@@ -8,6 +8,7 @@ use Illuminate\Support\ServiceProvider;
 use Liberu\ControlPanel\Mail\Actions\ConfigureMailControls;
 use Liberu\ControlPanel\Mail\Actions\CreateMailAccount;
 use Liberu\ControlPanel\Mail\Actions\CreateMailAlias;
+use Liberu\ControlPanel\Mail\Actions\CreateMailRoute;
 use Liberu\ControlPanel\Mail\Actions\RecordDeliveryDiagnostic;
 use Liberu\ControlPanel\Mail\Actions\RecordMailOperation;
 use Liberu\ControlPanel\Mail\Actions\RegisterMailDomain;
@@ -20,6 +21,7 @@ final class MailServiceProvider extends ServiceProvider
     {
         $this->app->scoped(CreateMailAccount::class);
         $this->app->scoped(CreateMailAlias::class);
+        $this->app->scoped(CreateMailRoute::class);
         $this->app->scoped(ConfigureMailControls::class);
         $this->app->scoped(RecordDeliveryDiagnostic::class);
         $this->app->scoped(ListMailAccounts::class);